move chat completions to a separate component
