--- src/pages/conversation/LoadingRing.css
@@ -1,36 +1,0 @@
-.loading-ring {
-  display: inline-block;
-  position: relative;
-  width: 80px;
-  height: 80px;
-  margin-top: 20px;
-}
-.loading-ring div {
-  box-sizing: border-box;
-  display: block;
-  position: absolute;
-  width: 64px;
-  height: 64px;
-  margin: 8px;
-  border: 8px solid #fff;
-  border-radius: 50%;
-  animation: lds-ring 1.2s cubic-bezier(0.5, 0, 0.5, 1) infinite;
-  border-color: #fff transparent transparent transparent;
-}
-.loading-ring div:nth-child(1) {
-  animation-delay: -0.45s;
-}
-.loading-ring div:nth-child(2) {
-  animation-delay: -0.3s;
-}
-.loading-ring div:nth-child(3) {
-  animation-delay: -0.15s;
-}
-@keyframes lds-ring {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
